message: Split MessageDict and friends into its own file.
(cherry picked from commit 93a071a1f8050c7b99ea4b3b34bc6a290db77f28)
