+++ zerver/tests/test_reactions.py
@@ -9,7 +9,7 @@ from zerver.actions.streams import do_change_stream_permission
 from zerver.lib.cache import cache_get, to_dict_cache_key_id
 from zerver.lib.emoji import get_emoji_data
 from zerver.lib.exceptions import JsonableError
-from zerver.lib.message import extract_message_dict
+from zerver.lib.message_cache import extract_message_dict
 from zerver.lib.test_classes import ZulipTestCase
 from zerver.lib.test_helpers import zulip_reaction_info
 from zerver.models import Message, Reaction, RealmEmoji, UserMessage
